feat: organize about-us and home into modules folder
