+++ src/modules/about-us/OurMission/index.ts
@@ -1,0 +1,1 @@
+export { default as OurMission } from "./OurMission";
